hanoip: Initiate dual patch concept for touchscreen universal fix
As usually we thought everything is settled with previous commits but unfortunately touch broke with custom inline kernel. No idea what really went wrong but according to dmesg we identified that signature was not getting matched with the kernel we use to load modules. Especially when we load modules from stock vendor then it has stock signature and inline kernel with custom kernel  signature. So only way for us to settle under permanent bind for temporary boot via firmware and modules from stock but we don't load them because it works without loading. Coming to second patch this will divert the path to sbin for stock users. Because on stock rom when we use twrp as permanent then we cannot sync stock kernel with different modules Because again signature miss match. So we will keep inline kernel modules in vendor but we don't load them ( this will resolve for temporary boot ) and for stock rom users twrp second patch will pick ongoing rom modules & firmware via mounting vendor and push to sbin and load the modules via script. So that permanent zip on stock will be retained with same logic as planned via ongoing rom kernel ,firmware & kernel modules.

Signed-off-by: RaghuVarma331 <raghuvarma331@gmail.com>